+++ src/main/java/dev/gaycookie/geodes/GeodeLootTableEntry.java
@@ -1,0 +1,11 @@
+package dev.gaycookie.geodes;
+
+public class GeodeLootTableEntry {
+  public final double weight;
+  public final String item;
+
+  public GeodeLootTableEntry(double weight, String item) {
+      this.weight = weight;
+      this.item = item;
+  }
+}